requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
playwright
openai
requests